Move Analyzer to ILSpyX (temporary: #nullable disable)
