+++ ICSharpCode.ILSpyX/Analyzers/Builtin/FindTypeInAttributeDecoder.cs
@@ -25,7 +25,7 @@ using ICSharpCode.Decompiler;
 using ICSharpCode.Decompiler.Metadata;
 using ICSharpCode.Decompiler.TypeSystem;
 
-namespace ICSharpCode.ILSpy.Analyzers.Builtin
+namespace ICSharpCode.ILSpyX.Analyzers.Builtin
 {
 	public enum TokenSearchResult : byte
 	{
