+++ ICSharpCode.ILSpyX/Analyzers/Builtin/MethodImplementedByAnalyzer.cs
@@ -16,33 +16,38 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
+#nullable disable
+
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.Diagnostics;
 using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
 using ICSharpCode.Decompiler.TypeSystem;
 
-namespace ICSharpCode.ILSpy.Analyzers.Builtin
+namespace ICSharpCode.ILSpyX.Analyzers.Builtin
 {
 	/// <summary>
-	/// Shows properties that implement an interface property.
+	/// Shows methods that implement an interface method.
 	/// </summary>
-	[ExportAnalyzer(Header = "Implemented By", Order = 10)]
-	class PropertyImplementedByAnalyzer : IAnalyzer
+	[ExportAnalyzer(Header = "Implemented By", Order = 40)]
+	class MethodImplementedByAnalyzer : IAnalyzer
 	{
 		public IEnumerable<ISymbol> Analyze(ISymbol analyzedSymbol, AnalyzerContext context)
 		{
-			Debug.Assert(analyzedSymbol is IProperty);
-			var scope = context.GetScopeOf((IProperty)analyzedSymbol);
+			Debug.Assert(analyzedSymbol is IMethod);
+			var scope = context.GetScopeOf((IEntity)analyzedSymbol);
 			foreach (var type in scope.GetTypesInScope(context.CancellationToken))
 			{
-				foreach (var result in AnalyzeType((IProperty)analyzedSymbol, type))
+				foreach (var result in AnalyzeType((IMethod)analyzedSymbol, type))
 					yield return result;
 			}
 		}
 
-		IEnumerable<IEntity> AnalyzeType(IProperty analyzedEntity, ITypeDefinition type)
+		IEnumerable<IEntity> AnalyzeType(IMethod analyzedEntity, ITypeDefinition type)
 		{
 			var token = analyzedEntity.MetadataToken;
 			var declaringTypeToken = analyzedEntity.DeclaringTypeDefinition.MetadataToken;
@@ -51,17 +56,17 @@ namespace ICSharpCode.ILSpy.Analyzers.Builtin
 			if (!allTypes.Any(t => t.MetadataToken == declaringTypeToken && t.ParentModule.PEFile == module))
 				yield break;
 
-			foreach (var property in type.Properties)
+			foreach (var method in type.Methods)
 			{
-				var baseMembers = InheritanceHelper.GetBaseMembers(property, true);
+				var baseMembers = InheritanceHelper.GetBaseMembers(method, true);
 				if (baseMembers.Any(m => m.MetadataToken == token && m.ParentModule.PEFile == module))
-					yield return property;
+					yield return method;
 			}
 		}
 
-		public bool Show(ISymbol symbol)
+		public bool Show(ISymbol entity)
 		{
-			return symbol is IProperty entity && entity.DeclaringType.Kind == TypeKind.Interface;
+			return entity is IMethod method && method.DeclaringType.Kind == TypeKind.Interface;
 		}
 	}
 }
