+++ ICSharpCode.ILSpyX/Analyzers/Builtin/MethodUsedByAnalyzer.cs
@@ -16,31 +16,38 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
+#nullable disable
+
+using System;
 using System.Collections.Generic;
+using System.ComponentModel.Composition;
 using System.Diagnostics;
 using System.Linq;
 using System.Reflection.Metadata;
 
+using ICSharpCode.Decompiler;
 using ICSharpCode.Decompiler.Disassembler;
 using ICSharpCode.Decompiler.Metadata;
 using ICSharpCode.Decompiler.TypeSystem;
 
-namespace ICSharpCode.ILSpy.Analyzers.Builtin
+namespace ICSharpCode.ILSpyX.Analyzers.Builtin
 {
 	/// <summary>
 	/// Shows entities that are used by a method.
 	/// </summary>
 	[ExportAnalyzer(Header = "Used By", Order = 20)]
-	class MethodVirtualUsedByAnalyzer : IAnalyzer
+	class MethodUsedByAnalyzer : IAnalyzer
 	{
 		const GetMemberOptions Options = GetMemberOptions.IgnoreInheritedMembers | GetMemberOptions.ReturnMemberDefinitions;
 
-		public bool Show(ISymbol symbol) => symbol is IMethod method && method.IsVirtual;
+		public bool Show(ISymbol symbol) => symbol is IMethod method && !method.IsVirtual && method.ParentModule is not null;
 
 		public IEnumerable<ISymbol> Analyze(ISymbol analyzedSymbol, AnalyzerContext context)
 		{
 			Debug.Assert(analyzedSymbol is IMethod);
+
 			var analyzedMethod = (IMethod)analyzedSymbol;
+			var analyzedBaseMethod = (IMethod)InheritanceHelper.GetBaseMember(analyzedMethod);
 			var mapping = context.Language
 				.GetCodeMappingInfo(analyzedMethod.ParentModule.PEFile,
 					analyzedMethod.DeclaringTypeDefinition.MetadataToken);
@@ -53,12 +60,13 @@ namespace ICSharpCode.ILSpy.Analyzers.Builtin
 			foreach (var type in scope.GetTypesInScope(context.CancellationToken))
 			{
 				var parentModule = (MetadataModule)type.ParentModule;
-				mapping = context.Language.GetCodeMappingInfo(parentModule.PEFile, type.MetadataToken);
+				mapping = null;
 				var methods = type.GetMembers(m => m is IMethod, Options).OfType<IMethod>();
 				foreach (var method in methods)
 				{
-					if (IsUsedInMethod((IMethod)analyzedSymbol, method, context))
+					if (IsUsedInMethod(analyzedMethod, analyzedBaseMethod, method, context))
 					{
+						mapping ??= context.Language.GetCodeMappingInfo(parentModule.PEFile, type.MetadataToken);
 						var parent = mapping.GetParentMethod((MethodDefinitionHandle)method.MetadataToken);
 						yield return parentModule.GetDefinition(parent);
 					}
@@ -66,12 +74,12 @@ namespace ICSharpCode.ILSpy.Analyzers.Builtin
 
 				foreach (var property in type.Properties)
 				{
-					if (property.CanGet && IsUsedInMethod((IMethod)analyzedSymbol, property.Getter, context))
+					if (property.CanGet && IsUsedInMethod(analyzedMethod, analyzedBaseMethod, property.Getter, context))
 					{
 						yield return property;
 						continue;
 					}
-					if (property.CanSet && IsUsedInMethod((IMethod)analyzedSymbol, property.Setter, context))
+					if (property.CanSet && IsUsedInMethod(analyzedMethod, analyzedBaseMethod, property.Setter, context))
 					{
 						yield return property;
 						continue;
@@ -80,71 +88,113 @@ namespace ICSharpCode.ILSpy.Analyzers.Builtin
 
 				foreach (var @event in type.Events)
 				{
-					if (@event.CanAdd && IsUsedInMethod((IMethod)analyzedSymbol, @event.AddAccessor, context))
+					if (@event.CanAdd && IsUsedInMethod(analyzedMethod, analyzedBaseMethod, @event.AddAccessor, context))
 					{
 						yield return @event;
 						continue;
 					}
-					if (@event.CanRemove && IsUsedInMethod((IMethod)analyzedSymbol, @event.RemoveAccessor, context))
+					if (@event.CanRemove && IsUsedInMethod(analyzedMethod, analyzedBaseMethod, @event.RemoveAccessor, context))
 					{
 						yield return @event;
 						continue;
 					}
-					if (@event.CanInvoke && IsUsedInMethod((IMethod)analyzedSymbol, @event.InvokeAccessor, context))
+					if (@event.CanInvoke && IsUsedInMethod(analyzedMethod, analyzedBaseMethod, @event.InvokeAccessor, context))
 					{
 						yield return @event;
 						continue;
 					}
 				}
-
 			}
 		}
 
-		bool IsUsedInMethod(IMethod analyzedEntity, IMethod method, AnalyzerContext context)
+		bool IsUsedInMethod(IMethod analyzedEntity, IMethod analyzedBaseMethod, IMethod method, AnalyzerContext context)
 		{
-			return ScanMethodBody(analyzedEntity, method, context.GetMethodBody(method));
+			return ScanMethodBody(analyzedEntity, method, analyzedBaseMethod, context.GetMethodBody(method));
 		}
 
-		static bool ScanMethodBody(IMethod analyzedMethod, IMethod method, MethodBodyBlock methodBody)
+		static bool ScanMethodBody(IMethod analyzedMethod, IMethod method, IMethod analyzedBaseMethod, MethodBodyBlock methodBody)
 		{
 			if (methodBody == null)
 				return false;
+
 			var mainModule = (MetadataModule)method.ParentModule;
 			var blob = methodBody.GetILReader();
-			var genericContext = new Decompiler.TypeSystem.GenericContext();
+
+			var genericContext = new Decompiler.TypeSystem.GenericContext(); // type parameters don't matter for this analyzer
 
 			while (blob.RemainingBytes > 0)
 			{
-				var opCode = blob.DecodeOpCode();
-				switch (opCode.GetOperandType())
+				ILOpCode opCode;
+				try
 				{
-					case OperandType.Method:
-					case OperandType.Sig:
-					case OperandType.Tok:
-						var member = MetadataTokenHelpers.EntityHandleOrNil(blob.ReadInt32());
-						if (member.IsNil)
-							continue;
-
-						switch (member.Kind)
-						{
-							case HandleKind.MethodDefinition:
-							case HandleKind.MethodSpecification:
-							case HandleKind.MemberReference:
-								var m = (mainModule.ResolveEntity(member, genericContext) as IMember)?.MemberDefinition;
-								if (m != null && m.MetadataToken == analyzedMethod.MetadataToken && m.ParentModule.PEFile == analyzedMethod.ParentModule.PEFile)
-								{
-									return true;
-								}
-								break;
-						}
-						break;
-					default:
+					opCode = blob.DecodeOpCode();
+					if (!IsSupportedOpCode(opCode))
+					{
 						ILParser.SkipOperand(ref blob, opCode);
-						break;
+						continue;
+					}
+				}
+				catch (BadImageFormatException)
+				{
+					return false; // unexpected end of blob
+				}
+				var member = MetadataTokenHelpers.EntityHandleOrNil(blob.ReadInt32());
+				if (!AnalyzerHelpers.IsPossibleReferenceTo(member, mainModule.PEFile, analyzedMethod))
+				{
+					if (analyzedBaseMethod == null || !AnalyzerHelpers.IsPossibleReferenceTo(member, mainModule.PEFile, analyzedBaseMethod))
+					{
+						continue;
+					}
+				}
+
+				IMember m;
+				try
+				{
+					m = (mainModule.ResolveEntity(member, genericContext) as IMember)?.MemberDefinition;
+				}
+				catch (BadImageFormatException)
+				{
+					continue;
+				}
+				if (m == null)
+					continue;
+
+				if (opCode == ILOpCode.Callvirt && analyzedBaseMethod != null)
+				{
+					if (IsSameMember(analyzedBaseMethod, m))
+					{
+						return true;
+					}
+				}
+				if (IsSameMember(analyzedMethod, m))
+				{
+					return true;
 				}
 			}
 
 			return false;
+		}
+
+		static bool IsSupportedOpCode(ILOpCode opCode)
+		{
+			switch (opCode)
+			{
+				case ILOpCode.Call:
+				case ILOpCode.Callvirt:
+				case ILOpCode.Ldtoken:
+				case ILOpCode.Ldftn:
+				case ILOpCode.Ldvirtftn:
+				case ILOpCode.Newobj:
+					return true;
+				default:
+					return false;
+			}
+		}
+
+		static bool IsSameMember(IMember analyzedMethod, IMember m)
+		{
+			return m.MetadataToken == analyzedMethod.MetadataToken
+				&& m.ParentModule.PEFile == analyzedMethod.ParentModule.PEFile;
 		}
 	}
 }
